Create dest if not exists
The current one does not create it and relies on the creation in one of
the item if the item is a dir using os.MkDirALl however it is subject to
a race condition when the first item ever is a file it will return
error.

This simple fix is just create the root destination dir.